+++ src/main/scala/dxWDL/Utils.scala
@@ -102,7 +102,6 @@ object Utils {
     val INSTANCE_TYPE_DB_FILENAME = "instanceTypeDB.json"
     val INTERMEDIATE_RESULTS_FOLDER = "intermediate"
     val LINK_INFO_FILENAME = "linking.json"
-    val MAX_HOURLY_RATE = 10.0
     val MAX_STRING_LEN = 8 * 1024     // Long strings cause problems with bash and the UI
     val MAX_NUM_FILES_MOVE_LIMIT = 1000
     val OUTPUT_SECTION = "outputs"
@@ -121,7 +120,7 @@ object Utils {
     val reservedSubstrings = List("___")
 
     // Prefixes used for generated applets
-    val reservedAppletPrefixes = List(SCATTER, COMMON)
+    val reservedAppletPrefixes = List(SCATTER, IF)
 
     lazy val execDirPath : Path = {
         val currentDir = System.getProperty("user.dir")